requirements.txt
Django
beautifulsoup4
requests
django-admin-rangefilter
psycopg2
